Optional dependency and some package restructuring
