+++ src/main/java/org/jongo/MongoCursor.java
@@ -23,6 +23,7 @@ import java.io.Closeable;
 import java.io.IOException;
 import java.util.Iterator;
 import java.util.NoSuchElementException;
+import org.jongo.function.Function;
 
 public class MongoCursor<E> implements Iterator<E>, Iterable<E>, Closeable {
